+++ tests/test_util.py
@@ -1,10 +1,17 @@
 import pytest
 from sympl import (
     set_prognostic_update_frequency, Prognostic, replace_none_with_default,
-    default_constants, ensure_no_shared_keys, SharedKeyException)
+    default_constants, ensure_no_shared_keys, SharedKeyException, DataArray,
+    combine_dimensions, set_dimension_names)
 from sympl._core.util import update_dict_by_adding_another
 from datetime import datetime, timedelta
 import numpy as np
+import unittest
+
+
+def same_list(list1, list2):
+    return (len(list1) == len(list2) and all(
+        [item in list2 for item in list1] + [item in list1 for item in list2]))
 
 
 class MockPrognostic(Prognostic):
@@ -165,6 +172,62 @@ def test_ensure_no_shared_keys_with_shared_keys():
         raise AssertionError(
             'No exception raised but expected SharedKeyException.')
 
+
+class CombineDimensionsTests(unittest.TestCase):
+
+    def setUp(self):
+        self.array_1d = DataArray(np.zeros((2,)), dims=['lon'])
+        self.array_2d = DataArray(np.zeros((2, 2)), dims=['lat', 'lon'])
+        self.array_3d = DataArray(np.zeros((2, 2, 2)),
+                                  dims=['lon', 'lat', 'interface_levels'])
+        set_dimension_names(
+            x=['lon'], y=['lat'], z=['mid_levels', 'interface_levels'])
+
+    def tearDown(self):
+        set_dimension_names(x=[], y=[], z=[])
+
+    def test_combine_dimensions_2d_and_3d(self):
+        dims = combine_dimensions(
+            [self.array_2d, self.array_3d], out_dims=('x', 'y', 'z'))
+        assert same_list(dims, ['lon', 'lat', 'interface_levels'])
+
+    def test_combine_dimensions_2d_and_3d_z_y_x(self):
+        dims = combine_dimensions(
+            [self.array_2d, self.array_3d], out_dims=('z', 'y', 'x'))
+        assert same_list(dims, ['interface_levels', 'lat', 'lon'])
+
+    def combine_dimensions_1d_shared(self):
+        dims = combine_dimensions(
+            [self.array_1d, self.array_1d], out_dims=['x'])
+        assert same_list(dims, ['lon'])
+
+    def combine_dimensions_1d_not_shared(self):
+        array_1d_x = DataArray(np.zeros((2,)), dims=['lon'])
+        array_1d_y = DataArray(np.zeros((2,)), dims=['lat'])
+        dims = combine_dimensions([array_1d_x, array_1d_y], out_dims=['x', 'y'])
+        assert same_list(dims, ['lon', 'lat'])
+
+    def combine_dimensions_1d_wrong_direction(self):
+        try:
+            combine_dimensions(
+                [self.array_1d, self.array_1d], out_dims=['z'])
+        except ValueError:
+            pass
+        except Exception as err:
+            raise err
+        else:
+            raise AssertionError('No exception raised but expected ValueError.')
+
+    def combine_dimensions_1d_and_2d_extra_direction(self):
+        try:
+            combine_dimensions(
+                [self.array_1d, self.array_2d], out_dims=['y'])
+        except ValueError:
+            pass
+        except Exception as err:
+            raise err
+        else:
+            raise AssertionError('No exception raised but expected ValueError.')
 
 if __name__ == '__main__':
     pytest.main([__file__])